+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_repr.py
@@ -351,24 +351,7 @@ def document_formula(bounding_box, document_span):
     return model, model_repr
 
 @pytest.fixture
-def document_image(bounding_box, document_span):
-    model = _models.DocumentImage(
-                    page_number=1,
-                    polygon=bounding_box[0],
-                    span=document_span[0],
-                    confidence=0.8
-                )
-    model_repr = "DocumentImage(page_number={}, polygon={}, confidence={}, span={})".format(
-            1,
-            bounding_box[1],
-            0.8,
-            document_span[1],
-        )
-    assert repr(model) == model_repr
-    return model, model_repr
-
-@pytest.fixture
-def document_page(document_span, document_word, document_selection_mark, document_line, document_annotation, document_formula, document_image, document_barcode):
+def document_page(document_span, document_word, document_selection_mark, document_line, document_annotation, document_formula, document_barcode):
     model = _models.DocumentPage(
         page_number=1,
         angle=120.0,
@@ -382,11 +365,10 @@ def document_page(document_span, document_word, document_selection_mark, documen
         kind="document",
         annotations=[document_annotation[0]],
         formulas=[document_formula[0]],
-        images=[document_image[0]],
         barcodes=[document_barcode[0]],
     )
     model_repr = "DocumentPage(page_number={}, angle={}, width={}, height={}, unit={}, lines=[{}], words=[{}], selection_marks=[{}], spans=[{}], " \
-                 "kind={}, annotations=[{}], barcodes=[{}], formulas=[{}], images=[{}])".format(
+                 "kind={}, annotations=[{}], barcodes=[{}], formulas=[{}])".format(
                 1,
                 120.0,
                 8.0,
@@ -400,7 +382,6 @@ def document_page(document_span, document_word, document_selection_mark, documen
                 document_annotation[1],
                 document_barcode[1],
                 document_formula[1],
-                document_image[1],
             )
     assert repr(model) == model_repr
     return model, model_repr